refactor: shared 훅 이동
