+++ src/basic/hooks/index.ts
@@ -1,10 +1,6 @@
-// Domain-specific hooks
 export * from './product';
 export * from './cart';
 export * from './coupon';
 export * from './admin';
 export * from './search';
 export * from './notification';
-
-// Utility hooks
-export * from './utils';
